refactor: moving dns/eventgrid
