+++ azurerm/internal/services/eventgrid/client/client.go
@@ -1,4 +1,4 @@
-package eventgrid
+package client
 
 import (
 	"github.com/Azure/azure-sdk-for-go/services/preview/eventgrid/mgmt/2018-09-15-preview/eventgrid"
@@ -11,7 +11,7 @@ type Client struct {
 	TopicsClient             *eventgrid.TopicsClient
 }
 
-func BuildClient(o *common.ClientOptions) *Client {
+func NewClient(o *common.ClientOptions) *Client {
 	DomainsClient := eventgrid.NewDomainsClientWithBaseURI(o.ResourceManagerEndpoint, o.SubscriptionId)
 	o.ConfigureClient(&DomainsClient.Client, o.ResourceManagerAuthorizer)
